Add highlighting on modal
